# app/contracts/src/generated/auth/src/main/java/com/luckydraw/contracts/auth/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.luckydraw.contracts.auth.api;

import com.luckydraw.contracts.auth.api.model.ErrorEnvelope;
import com.luckydraw.contracts.auth.api.model.GetAuthJwksResponseDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthLoginRequestDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthLoginResponseDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthLogoutResponseDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthRefreshRequestDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthRefreshResponseDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthRegisterRequestDTO;
import com.luckydraw.contracts.auth.api.model.PostAuthRegisterResponseDTO;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.9.0")
@Validated
public interface AuthApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /auth/.well-known/jwks.json : 取得驗證公鑰（JWKS） [auth-keys-001]
     * 公開簽章公鑰（JWKS，RFC 7517），供 Gateway 與各服務**獨立驗證** JWT 簽章（ADR-009 defense in depth）。 不需登入（PUBLIC，&#x60;FR-GW-06&#x60; / AC-AUTH-008）。  - 支援多 key 輪替（&#x60;keys&#x60; 陣列）；RSA key 含 &#x60;kty&#x60;/&#x60;kid&#x60;/&#x60;n&#x60;/&#x60;e&#x60;/&#x60;alg&#x60;（及 &#x60;use&#x60;）。 - 僅公開 **public key**，簽發 private key 存放於 GCP Secret Manager（ADR-008），**無法經任何端點取得**（AC-AUTH-007）。  **Traceability:** &#x60;FR-AUTH-03&#x60; / &#x60;AC-AUTH-007 ~ 008&#x60; 
     *
     * @return 成功取得 JWKS (status code 200)
     *         or 系統錯誤 → &#x60;B0000&#x60; (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/auth/.well-known/jwks.json",
        produces = { "application/json" }
    )
    
    default ResponseEntity<GetAuthJwksResponseDTO> getAuthJwks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"data\" : { \"keys\" : [ { \"kty\" : \"RSA\", \"use\" : \"sig\", \"alg\" : \"RS256\", \"kid\" : \"key-1\", \"n\" : \"0vx7agoebGcQSuuPiLJXZptN9nndrQmbXEps2aiAFbWhM78LhWx4cbbfAAtVT86zwu1RK7aPFFxuhDR1L6tSoc_BJECPebWKRXjBZCiFV4n3oknjhMstn64tZ_2W-5JsGY4Hc5n9yBXArwl93lqt7_RN5w6Cf0h4QyQ5v-65YGjQR0_FDW2QvzqY368QQMicAtaSqzs8KJZgnYb9c7d0zgdAZHzu6qMQvRL5hajrn1n91CbOpbISD08qNLyrdkt-bFTWhAI4vMQFh6WeZu0fM4lFd2NcRwr3XPksINHaQ-G_xBniIqbw0Ls1jF44-csFCur-kEgU8awapJzKnqDKgw\", \"e\" : \"AQAB\" } ] }, \"message\" : \"OK\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/login : 登入並簽發憑證 [auth-tokens-001]
     * 驗證憑證後簽發身份憑證（UC-2）。&#x60;username&#x60; 欄位可填 **username 或 email**（SA UC-2）。  - 成功：&#x60;200&#x60;，回傳 &#x60;{ accessToken, tokenType, expiresIn, refreshToken }&#x60;。   accessToken 為 RS256 JWT，claims 承載 &#x60;sub&#x60;（userId）、&#x60;roles&#x60;、&#x60;exp&#x60;、&#x60;iat&#x60;、&#x60;iss&#x60;（ADR-009）。 - 失敗：**「帳號不存在」與「密碼錯誤」回同一 &#x60;401&#x60; / &#x60;A0201&#x60;**，不洩漏帳號存在性（AC-AUTH-006）。 - 憑證承載內容（身份/角色/有效期/簽發時刻/簽發者）全由伺服端決定，client 不得指定（AC-AUTH-011）。  **Traceability:** &#x60;FR-AUTH-02&#x60;, &#x60;FR-AUTH-03&#x60;, &#x60;FR-X-01&#x60; / &#x60;AC-AUTH-004 ~ 006&#x60;, &#x60;AC-AUTH-012&#x60; 
     *
     * @param postAuthLoginRequestDTO  (required)
     * @return 登入成功，簽發身份憑證（含 refreshToken，Should） (status code 200)
     *         or 結構性輸入錯誤（username/password 缺欄位或空白）→ &#x60;A0000&#x60; (status code 400)
     *         or 登入失敗：**帳號不存在與密碼錯誤回同一碼 &#x60;A0201&#x60;**（不洩漏存在性）。 二級碼 &#x60;A0200&#x60; 作為登入異常的通用碼。  (status code 401)
     *         or 系統錯誤 → &#x60;B0000&#x60; (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PostAuthLoginResponseDTO> postAuthLogin(
         @Valid @RequestBody PostAuthLoginRequestDTO postAuthLoginRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"data\" : { \"accessToken\" : \"eyJhbGciOiJSUzI1NiIsImtpZCI6ImtleS0xIn0...\", \"tokenType\" : \"Bearer\", \"expiresIn\" : 1800, \"refreshToken\" : \"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...\" }, \"message\" : \"OK\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/logout : 登出並撤銷憑證 [auth-tokens-003]
     * 登出（撤銷）當前 bearer token（ADR-009 修訂：token 白名單）。將 token 的 jti 移出白名單， 之後該 token 即使簽章有效、未過期，也會因「不在白名單」而被 Gateway/各服務拒絕（REVOKED → &#x60;401&#x60;/&#x60;A0203&#x60;）。  - 需持有效 bearer token（受保護，&#x60;Authorization: Bearer &lt;token&gt;&#x60;）。 - 成功：&#x60;200&#x60;，回傳空 &#x60;data&#x60;。登出為**最佳努力撤銷**（冪等）：token 缺失/已過期/已登出 → 仍回 &#x60;200&#x60;。 - 僅撤銷「當下」這個 token（jti 維度），不影響同使用者的其他 token。  **Traceability:** &#x60;FR-AUTH&#x60;（登出/撤銷，ADR-009 修訂） 
     *
     * @return 登出成功（撤銷 token） (status code 200)
     *         or 憑證無效/過期/已撤銷 → &#x60;A0203&#x60;（&#x60;A0202&#x60; 過期） (status code 401)
     *         or 系統錯誤 → &#x60;B0000&#x60; (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/logout",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PostAuthLogoutResponseDTO> postAuthLogout(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"data\" : \"{}\", \"message\" : \"OK\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/refresh : 刷新存取憑證 [auth-tokens-002] (Should)
     * 以刷新憑證換發新的存取憑證（UC-3，**Should**：prod 前完成，非 POC 必須，&#x60;FR-AUTH-04&#x60;）。  - refresh 僅**延長登入有效期**，**不改變身份或權限**（沿用原身份與角色，更新有效期）。 - 刷新憑證過期或無效 → &#x60;401&#x60;（&#x60;A0202&#x60; 過期 / &#x60;A0203&#x60; 無效），使用者須重新登入（AC-AUTH-014）。 - 本端點標記 USER 角色（需持有有效憑證語境）；刷新憑證由 request body 的 &#x60;refreshToken&#x60; 承載。  **Traceability:** &#x60;FR-AUTH-04&#x60; / &#x60;AC-AUTH-013 ~ 014&#x60; 
     *
     * @param postAuthRefreshRequestDTO  (required)
     * @return 刷新成功，回傳新的存取憑證（身份/角色不變） (status code 200)
     *         or 刷新憑證過期 → &#x60;A0202&#x60;；刷新憑證無效（簽章/格式）→ &#x60;A0203&#x60;。 二級碼 &#x60;A0200&#x60; 作為憑證異常的通用碼。  (status code 401)
     *         or 權限不足 → &#x60;A0400&#x60;（未持有有效憑證語境） (status code 403)
     *         or 系統錯誤 → &#x60;B0000&#x60; (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/refresh",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PostAuthRefreshResponseDTO> postAuthRefresh(
         @Valid @RequestBody PostAuthRefreshRequestDTO postAuthRefreshRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"data\" : { \"accessToken\" : \"eyJhbGciOiJSUzI1NiIsImtpZCI6ImtleS0xIn0...\", \"tokenType\" : \"Bearer\", \"expiresIn\" : 1800 }, \"message\" : \"OK\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/register : 使用者註冊 [auth-users-001]
     * 建立使用者帳號（UC-1）。預設角色 &#x60;ROLE_USER&#x60;，密碼以不可逆方式雜湊儲存 （&#x60;FR-AUTH-06&#x60;）；**任何環節不得明文**。使用者不得於註冊時指定或自選角色。  - 成功：&#x60;200&#x60;，回傳使用者識別與角色，**絕不含 password / password_hash**（AC-AUTH-001）。 - 重複 username / email：&#x60;409&#x60;（&#x60;A0101&#x60; / &#x60;A0102&#x60;），不建立第二個帳號（AC-AUTH-002）。 - 輸入驗證失敗（缺欄位 / email 格式非法 / 密碼為空）：&#x60;400&#x60;（&#x60;A0103&#x60;），不建立帳號（AC-AUTH-003）。  **Traceability:** &#x60;FR-AUTH-01&#x60;, &#x60;FR-AUTH-06&#x60;, &#x60;FR-X-01&#x60; / &#x60;AC-AUTH-001 ~ 003&#x60; 
     *
     * @param postAuthRegisterRequestDTO  (required)
     * @return 註冊成功，回傳使用者識別與角色（不含密碼／雜湊） (status code 200)
     *         or 註冊輸入驗證失敗（缺欄位、email 格式非法、密碼為空）→ &#x60;A0103&#x60;。 一般註冊錯誤亦可用二級碼 &#x60;A0100&#x60;。  (status code 400)
     *         or 唯一性衝突：username 已存在 → &#x60;A0101&#x60;；email 已存在 → &#x60;A0102&#x60;。 二級碼 &#x60;A0100&#x60; 亦可作為通用註冊衝突。  (status code 409)
     *         or 系統錯誤 → &#x60;B0000&#x60;（雜湊/儲存異常等，不得產生半完成資料） (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/register",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PostAuthRegisterResponseDTO> postAuthRegister(
         @Valid @RequestBody PostAuthRegisterRequestDTO postAuthRegisterRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"data\" : { \"id\" : 1001, \"username\" : \"alice\", \"email\" : \"alice@example.com\", \"roles\" : [ \"ROLE_USER\" ] }, \"message\" : \"OK\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0201\", \"message\" : \"帳號或密碼錯誤\", \"data\" : null }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}